+++ pkg/animal/birds.go
@@ -1,5 +1,7 @@
 package animal
 
+import "github.com/ironarachne/world/pkg/size"
+
 func getBirds() []Animal {
 	birds := []Animal{
 		Animal{
@@ -25,7 +27,7 @@ func getBirds() []Animal {
 			MaxHumidity:    10,
 			MinTemperature: 0,
 			MaxTemperature: 10,
-			Size:           "tiny",
+			Size:           size.GetCategoryByName("tiny"),
 		},
 		Animal{
 			Name:           "chicken",
@@ -50,7 +52,7 @@ func getBirds() []Animal {
 			MaxHumidity:    10,
 			MinTemperature: 0,
 			MaxTemperature: 10,
-			Size:           "tiny",
+			Size:           size.GetCategoryByName("tiny"),
 		},
 		Animal{
 			Name:           "mudhen",
@@ -75,7 +77,7 @@ func getBirds() []Animal {
 			MaxHumidity:    10,
 			MinTemperature: 0,
 			MaxTemperature: 10,
-			Size:           "tiny",
+			Size:           size.GetCategoryByName("tiny"),
 		},
 		Animal{
 			Name:           "quail",
@@ -100,7 +102,7 @@ func getBirds() []Animal {
 			MaxHumidity:    10,
 			MinTemperature: 0,
 			MaxTemperature: 10,
-			Size:           "tiny",
+			Size:           size.GetCategoryByName("tiny"),
 		},
 		Animal{
 			Name:           "duck",
@@ -125,7 +127,7 @@ func getBirds() []Animal {
 			MaxHumidity:    10,
 			MinTemperature: 0,
 			MaxTemperature: 10,
-			Size:           "tiny",
+			Size:           size.GetCategoryByName("tiny"),
 		},
 		Animal{
 			Name:           "flamingo",
@@ -150,7 +152,7 @@ func getBirds() []Animal {
 			MaxHumidity:    10,
 			MinTemperature: 5,
 			MaxTemperature: 10,
-			Size:           "tiny",
+			Size:           size.GetCategoryByName("tiny"),
 		},
 		Animal{
 			Name:           "eagle",
@@ -175,7 +177,7 @@ func getBirds() []Animal {
 			MaxHumidity:    10,
 			MinTemperature: 0,
 			MaxTemperature: 10,
-			Size:           "tiny",
+			Size:           size.GetCategoryByName("tiny"),
 		},
 		Animal{
 			Name:           "falcon",
@@ -200,7 +202,7 @@ func getBirds() []Animal {
 			MaxHumidity:    10,
 			MinTemperature: 0,
 			MaxTemperature: 10,
-			Size:           "tiny",
+			Size:           size.GetCategoryByName("tiny"),
 		},
 		Animal{
 			Name:           "hawk",
@@ -225,7 +227,7 @@ func getBirds() []Animal {
 			MaxHumidity:    10,
 			MinTemperature: 0,
 			MaxTemperature: 10,
-			Size:           "tiny",
+			Size:           size.GetCategoryByName("tiny"),
 		},
 	}